Add scripts to support CUDA
CUDA itself is not shipped by EESSI and has to be installed on
the host. The scripts perform various checks and download and
install the CUDA compat libs.
Modules with CUDA as a dependecy are hidden in Lmod, unless
the CUDA compat libs are installed which is only done when CUDA
itself is installed on the host. This aspect still has to be
tested with an updated Lmod version in the EESSI compat layer.