Prepare shipping CUDA module file with EESSI
CUDA can now be installed in EESSI with the actual software being
installed in ${EESSI_SOFTWARE_PATH/versions/host_injections} and
the module file being written to ${EESSI_SOFTWARE_PATH}.
Since the host_injections directory is not shipped with EESSI,
the actual software will not be shipped. Trying to load the CUDA
module will fail with a pointer to the CUDA install script,
unless the software under host_injections is found.
The CUDA install script then installs the software in
host_injections, while the module file is written to a tmpdir
which is removed in the end.