Ship whitelisted CUDA libs and rework scripts accordingly
Whitelisted CUDA libraries can now be shipped with EESSI. The other
libraries and files are replaced with symlinks to host_injections.
A compiled CUDA sample can now also be shipped with EESSI. This is
relevant if users only need the runtime capabilities and not the
whole CUDA suite (which would include the compilers). It is now
possible to solely install the compat libs as a user and get access
to the runtime environment this way. It is still possible to also
install the whole CUDA suite.
CUDA enabled modules with the gpu property now only load if the
compat libs are installed in host_injections.