Add check if CUDA compat lib version is sufficient for module
The CUDA version needed for modules are now written as envvars
that will be exported into the module files. The CUDA version for
which we have the current compat libs installed is saved in a txt
file in ../host_injections/nvidia/latest/version.txt
The lmod hook called when loading modules with the gpu property
now compares these two versions and exits out if the installed
version needs to be updated.