{2023.06}[NVHPC/25.1-CUDA-12.6] add hook for nvhpc
This adds a pre_configure_hook for NVHPC. It performs some search and replace
operations on the "localrc" file used by NVHPC to detect information about the
system. In particular it points the sysroot flag at the eessi eprefix variable,
and appends two variables definitions about where to look for system libraries.

The content of the hook is extracted from:
https://github.com/ComputeCanada/easybuild-computecanada-config/blob/main/2023/cc_hooks.py#L544-L547